+++ scandeval/benchmarks/abstract/__init__.py
@@ -1,3 +1,6 @@
 from .base import BaseBenchmark  # noqa
 from .text_classification import TextClassificationBenchmark  # noqa
 from .token_classification import TokenClassificationBenchmark  # noqa
+from .dep import DepBenchmark  # noqa
+from .pos import PosBenchmark  # noqa
+from .ner import NerBenchmark  # noqa