+++ scandeval/benchmarks/ndt_nn_pos.py
@@ -1,19 +1,14 @@
 '''POS evaluation of a language model on the Nynorsk part of the NDT dataset'''
 
-from datasets import Dataset
-import numpy as np
-from typing import Tuple, Dict, List, Optional
 import logging
 
-from .abstract import TokenClassificationBenchmark
-from ..utils import doc_inherit
-from ..datasets import load_dataset
+from .abstract import PosBenchmark
 
 
 logger = logging.getLogger(__name__)
 
 
-class NdtNNPosBenchmark(TokenClassificationBenchmark):
+class NdtNNPosBenchmark(PosBenchmark):
     '''Benchmark of language models on the Nynorsk POS part of the NDT dataset.
 
     Args:
@@ -45,55 +40,7 @@ class NdtNNPosBenchmark(TokenClassificationBenchmark):
                  cache_dir: str = '.benchmark_models',
                  evaluate_train: bool = False,
                  verbose: bool = False):
-        id2label = ['ADJ', 'ADV', 'INTJ', 'NOUN', 'PROPN', 'VERB', 'ADP',
-                    'AUX', 'CCONJ', 'DET', 'NUM', 'PART', 'PRON', 'SCONJ',
-                    'PUNCT', 'SYM', 'X']
         super().__init__(name='ndt-nn-pos',
-                         metric_names=dict(accuracy='Accuracy'),
-                         id2label=id2label,
                          cache_dir=cache_dir,
                          evaluate_train=evaluate_train,
                          verbose=verbose)
-
-    @doc_inherit
-    def _load_data(self) -> Tuple[Dataset, Dataset]:
-        X_train, X_test, y_train, y_test = load_dataset(self.short_name)
-        train_dict = dict(doc=X_train['doc'],
-                          tokens=X_train['tokens'],
-                          orig_labels=y_train['pos_tags'])
-        test_dict = dict(doc=X_test['doc'],
-                         tokens=X_test['tokens'],
-                         orig_labels=y_test['pos_tags'])
-        train = Dataset.from_dict(train_dict)
-        test = Dataset.from_dict(test_dict)
-        return train, test
-
-    @doc_inherit
-    def _compute_metrics(self,
-                         predictions_and_labels: tuple,
-                         id2label: Optional[dict] = None) -> Dict[str, float]:
-        # Get the predictions from the model
-        predictions, labels = predictions_and_labels
-
-        if id2label is not None:
-            raw_predictions = np.argmax(predictions, axis=-1)
-
-            # Remove ignored index (special tokens)
-            predictions = [
-                [id2label[pred] for pred, lbl in zip(prediction, label)
-                 if lbl != -100]
-                for prediction, label in zip(raw_predictions, labels)
-            ]
-            labels = [
-                [id2label[lbl] for _, lbl in zip(prediction, label)
-                 if lbl != -100]
-                for prediction, label in zip(raw_predictions, labels)
-            ]
-
-        results = self._metric.compute(predictions=predictions,
-                                       references=labels)
-        return dict(accuracy=results['overall_accuracy'])
-
-    @doc_inherit
-    def _get_spacy_token_labels(self, processed) -> List[str]:
-        return [tok.pos_ for tok in processed]